backend/requirements.txt
django>=4.2
djangorestframework
django-cors-headers
qrcode
pillow
python-dotenv
